modularized firebase module
