+++ src/firebase/config.ts
@@ -1,0 +1,10 @@
+import { FirebaseOptions } from '@firebase/app-types';
+
+export const config: FirebaseOptions = {
+  apiKey: 'AIzaSyD1S_pjLMUUGFI78XC4cyyKRs9ba0rcktc',
+  authDomain: 'drivr-dev-4fa43.firebaseapp.com',
+  databaseURL: 'https://drivr-dev-4fa43.firebaseio.com',
+  projectId: 'drivr-dev-4fa43',
+  storageBucket: 'drivr-dev-4fa43.appspot.com',
+  messagingSenderId: '380534664070',
+};
