+++ src/firebase/init.ts
@@ -1,0 +1,8 @@
+import * as firebase from 'firebase/app';
+import { config } from './config';
+
+let initialized = false;
+if (!initialized) {
+  initialized = true;
+  firebase.initializeApp(config);
+}
